+++ test-scenario.sh
@@ -1,7 +1,3 @@
 #!/bin/bash
-tests=$1
-if [[ $tests = "" ]]; then
-  tests="all"
-fi
 
-java -jar lib/jstestdriver/JsTestDriver.jar --tests "$tests" --config jsTestDriver-scenario.conf --reset
+java -jar lib/jstestdriver/JsTestDriver.jar --tests "all" --config jsTestDriver-scenario.conf --reset